[not verified] rename hooks
